게시물 삭제 관련 백엔드 코드 변경

+++ src/main/java/PickMe/PickMeDemo/service/PostsService.java
@@ -353,33 +353,39 @@ public class PostsService {
 
 
     // 프로젝트 삭제
+    @EntityGraph(attributePaths = {"category"})
     public void deleteProject(Long projectId) {
         // projectId로 해당 게시물 찾기
         Posts posts = postsRepository.findById(projectId)
                 .orElseThrow(() -> new AppException("Project not found", HttpStatus.NOT_FOUND));
 
-        // projectId와 연관된 카테고리 찾기
-        Category category = categoryRepository.findCategoryById(projectId)
-                .orElseThrow(() -> new AppException("Category not found", HttpStatus.NOT_FOUND));
+        // 프로젝트 게시물과 연관된 카테고리 찾기
+        Category category = posts.getCategory();
 
-        // 카테고리 먼저 지워야
-        categoryRepository.delete(category);
+        if (category != null) { // 카테고리 먼저 지워야
+            // Delete the category
+            categoryRepository.delete(category);
+        }
+
         // 프로젝트도 삭제 가능
         postsRepository.delete(posts);
     }
 
     // 스터디 삭제
+    @EntityGraph(attributePaths = {"category"})
     public void deleteStudy(Long studyId) {
         // studyId로 해당 게시물 찾기
         Posts posts = postsRepository.findById(studyId)
                 .orElseThrow(() -> new AppException("Study not found", HttpStatus.NOT_FOUND));
 
-        // studyId와 연관된 카테고리 찾기
-        Category category = categoryRepository.findCategoryById(studyId)
-                .orElseThrow(() -> new AppException("Category not found", HttpStatus.NOT_FOUND));
+        // 스터디 게시물과 연관된 카테고리 찾기
+        Category category = posts.getCategory();
 
         // 카테고리 먼저 지워야
-        categoryRepository.delete(category);
+        if (category != null) { // 카테고리 먼저 지워야
+            // Delete the category
+            categoryRepository.delete(category);
+        }
         // 스터디도 삭제 가능
         postsRepository.delete(posts);
     }
